add selectRender() callback to the buttons and separate each functions to render into main
