+++ src/index.js
@@ -10,4 +10,3 @@ import { render } from './scripts/render';
 
 	return;
 })();
-
