+++ src/scripts/menu.js
@@ -1,0 +1,3 @@
+export default function makeMenu() {
+	
+}
