+++ src/styles/home.css
@@ -1,0 +1,44 @@
+.hero-container {
+	display: flex;
+	flex-direction: column-reverse;
+	justify-content: flex-end;
+
+	min-height: calc(100vh - 3em);
+
+	background-color: lightsteelblue;
+}
+
+.hero-text-container {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+
+	padding: 1em;
+}
+
+.hero-text-container h1 {
+	font-family: 'Merriweather', Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
+}
+
+.hero-cta {
+	padding: 1em;
+
+	background-color: yellow;
+	font-weight: bolder;
+}
+
+.hero-image {
+	width: 100%;
+}
+
+@media (min-width: 768px) {
+	.hero-container {
+		flex-direction: row;
+		justify-content: center;
+		align-items: center;
+		gap: 1em;
+	}
+	.hero-image-container {
+		width: 500px;
+	}
+}
