+++ src/styles/styles.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Merriweather:ital,wght@0,900;1,900&family=Roboto:wght@100;400;900&display=swap');
+
 /** {
 	outline: 1px dashed red;
 }*/
@@ -10,13 +12,7 @@ header {
 	display: block;
 }
 main {
-	display: flex;
-	flex-direction: column-reverse;
-	justify-content: flex-end;
-
-	min-height: calc(100vh - 3em);
-
-	background-color: lightsteelblue;
+	display: block;
 }
 footer {
 	padding: 1em;
@@ -32,7 +28,6 @@ footer {
 	padding-left: 0;
 
 	list-style: none;
-
 }
 
 .nav-link {
@@ -44,39 +39,3 @@ footer {
 	text-transform: uppercase;
 }
 
-.hero-text-container {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-
-	padding: 1em;
-}
-
-.hero-text-container h1 {
-	font-family: 'Merriweather', Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
-}
-
-.hero-cta {
-	padding: 1em;
-
-	background-color: yellow;
-	font-weight: bolder;
-}
-
-.hero-image {
-	width: 100%;
-}
-
-@import url('https://fonts.googleapis.com/css2?family=Merriweather:ital,wght@0,900;1,900&family=Roboto:wght@100;400;900&display=swap');
-
-@media (min-width: 768px) {
-	main {
-		flex-direction: row;
-		justify-content: center;
-		align-items: center;
-		gap: 1em;
-	}
-	.hero-image-container {
-		width: 500px;
-	}
-}
